feat: разделяет main.js на модули
